Return Ids for editorial purposes (#567)
* update name for garbo entity schema

* cleanup args and types

* add ids for editor

* improve seperation in schema folder

* move args above where clause

---------
