+++ src/api/services/emissionsService.ts
@@ -9,9 +9,10 @@ import {
   StatedTotalEmissions,
 } from '@prisma/client'
 import { OptionalNullable } from '../../lib/type-utils'
-import { DefaultEmissions, emissionsArgs } from '../types'
+import { DefaultEmissions } from '../types'
 import { prisma } from '../../lib/prisma'
 import { GarboAPIError } from '../../lib/garbo-api-error'
+import { emissionsArgs } from '../args'
 
 const TONNES_CO2_UNIT = 'tCO2e'
